--- internal/interfaces/router.go
@@ -1,5 +1,0 @@
-package interfaces
-
-type Router interface {
-	ConfigureRouter()
-}
