+++ internal/router/pinghandler.go
@@ -1,0 +1,25 @@
+package router
+
+import (
+	"encoding/json"
+	"net/http"
+)
+
+// Здесь создаем новый файл хэндлера и описываем свой хендлер
+func PingHandler(w http.ResponseWriter, r *http.Request) {
+	response := struct {
+		Message string `json:"message"`
+	}{
+		Message: "pong",
+	}
+
+	jsonResponse, err := json.Marshal(response)
+	if err != nil {
+		http.Error(w, err.Error(), http.StatusInternalServerError)
+	}
+
+	w.Header().Set("Content-Type", "application/json")
+
+	w.WriteHeader(http.StatusOK)
+	w.Write(jsonResponse)
+}
