--- internal/usecases/user/userReadCase.go
@@ -1,10 +1,0 @@
-package user
-
-import (
-	"AdHub/internal/entities"
-	"AdHub/internal/interfaces"
-)
-
-func UserRead(repo interfaces.UserRepo, login string) (*entities.User, error) {
-	return .Read(login)
-}
